# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Single sign-on (сервис единого входа)",
        "title": "SSO",
        "contact": {
            "name": "Alnovi",
            "url": "https://github.com/alnovi"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/alnovi/sso/refs/heads/master/LICENSE.md"
        },
        "version": "1.0.0"
    },
    "host": "http://localhost:8081",
    "paths": {
...
        "/api/sessions": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Список сессий пользователей",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Sessions"
                ],
                "summary": "Список сессий",
                "operationId": "SessionsList",
                "responses": {
                    "200": {
                        "description": "Информация о сессиях пользователей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.SessionUser"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/sessions/{id}": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Информация о сессии пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Sessions"
                ],
                "summary": "Сессия",
                "operationId": "SessionsGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор сессии",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о сессии пользователя",
                        "schema": {
                            "$ref": "#/definitions/response.SessionUser"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Удаление сессии пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Sessions"
                ],
                "summary": "Удаление сессии",
                "operationId": "SessionsDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор сессии",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/stats": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Метод возвращает количество пользователй, клиентов и сессий",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Stats"
                ],
                "summary": "Статистика",
                "operationId": "StatsAll",
                "responses": {
                    "200": {
                        "description": "Статистика",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Stats"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.SessionUser": {
            "type": "object",
            "properties": {
                "agent": {
                    "type": "string"
                },
                "app": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "is_current": {
                    "type": "boolean"
                },
                "os": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/response.User"
                }
            }
        },
        "response.Stats": {
            "type": "object",
            "properties": {
                "clients": {
                    "type": "integer"
                },
                "sessions": {
                    "type": "integer"
                },
                "users": {
                    "type": "integer"
                }
            }
        },
        "response.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT-Access": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}